requirements.txt
pytest==6.2.4
pytest-mock==3.6.1
PyYAML==5.4.1
requests==2.25.1
watchgod==0.7